♻️ Rework CC hook
