+++ apps/ledger-live-mobile/src/dynamicContent/useContentCards.ts
@@ -1,48 +1,8 @@
-import { useCallback, useEffect } from "react";
-
-import { useDispatch } from "react-redux";
-import {
-  setDynamicContentAssetsCards,
-  setDynamicContentWalletCards,
-  setDynamicContentNotificationCards,
-} from "../actions/dynamicContent";
-import { useBrazeContentCard } from "./brazeContentCard";
-import {
-  filterByPage,
-  mapAsWalletContentCard,
-  mapAsAssetContentCard,
-  mapAsNotificationContentCard,
-} from "./dynamicContent";
-import { LocationContentCard } from "./types";
+import { useEffect } from "react";
+import { useDynamicContentLogic } from "./useDynamicContentLogic";
 
 const HookDynamicContentCards = () => {
-  const dispatch = useDispatch();
-  const { Braze, refreshDynamicContent } = useBrazeContentCard();
-
-  const fetchData = useCallback(async () => {
-    // Fetch data from Braze
-    const contentCards = await Braze.getContentCards();
-
-    // Filtering v0
-    const walletCards = filterByPage(
-      contentCards,
-      LocationContentCard.Wallet,
-    ).map(card => mapAsWalletContentCard(card));
-
-    const assetCards = filterByPage(
-      contentCards,
-      LocationContentCard.Asset,
-    ).map(card => mapAsAssetContentCard(card));
-
-    const notificationCards = filterByPage(
-      contentCards,
-      LocationContentCard.NotificationCenter,
-    ).map(card => mapAsNotificationContentCard(card));
-
-    dispatch(setDynamicContentWalletCards(walletCards));
-    dispatch(setDynamicContentAssetsCards(assetCards));
-    dispatch(setDynamicContentNotificationCards(notificationCards));
-  }, [Braze, dispatch]);
+  const { refreshDynamicContent, fetchData } = useDynamicContentLogic();
 
   useEffect(() => {
     refreshDynamicContent();
