+++ apps/ledger-live-mobile/src/dynamicContent/useDynamicContentLogic.ts
@@ -1,0 +1,50 @@
+import { useCallback } from "react";
+import { useDispatch } from "react-redux";
+import {
+  setDynamicContentWalletCards,
+  setDynamicContentAssetsCards,
+  setDynamicContentNotificationCards,
+} from "../actions/dynamicContent";
+import { useBrazeContentCard } from "./brazeContentCard";
+import {
+  filterByPage,
+  mapAsWalletContentCard,
+  mapAsAssetContentCard,
+  mapAsNotificationContentCard,
+} from "./dynamicContent";
+import { LocationContentCard } from "./types";
+
+export const useDynamicContentLogic = () => {
+  const dispatch = useDispatch();
+  const { Braze, refreshDynamicContent } = useBrazeContentCard();
+
+  const fetchData = useCallback(async () => {
+    // Fetch data from Braze
+    const contentCards = await Braze.getContentCards();
+
+    // Filtering v0
+    const walletCards = filterByPage(
+      contentCards,
+      LocationContentCard.Wallet,
+    ).map(card => mapAsWalletContentCard(card));
+
+    const assetCards = filterByPage(
+      contentCards,
+      LocationContentCard.Asset,
+    ).map(card => mapAsAssetContentCard(card));
+
+    const notificationCards = filterByPage(
+      contentCards,
+      LocationContentCard.NotificationCenter,
+    ).map(card => mapAsNotificationContentCard(card));
+
+    dispatch(setDynamicContentWalletCards(walletCards));
+    dispatch(setDynamicContentAssetsCards(assetCards));
+    dispatch(setDynamicContentNotificationCards(notificationCards));
+  }, [Braze, dispatch]);
+
+  return {
+    refreshDynamicContent,
+    fetchData,
+  };
+};
